Combine two mocks
