+++ core/src/test/scala/org/broadinstitute/dsde/rawls/dataaccess/martha/MarthaResolverSpec.scala
@@ -2,12 +2,13 @@ package org.broadinstitute.dsde.rawls.dataaccess.martha
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.headers.OAuth2BearerToken
-import akka.stream.{ActorMaterializer, Materializer}
+import akka.stream.ActorMaterializer
 import org.broadinstitute.dsde.rawls.TestExecutionContext
+import org.broadinstitute.dsde.rawls.mock.MockMarthaResolver
 import org.broadinstitute.dsde.rawls.model.{RawlsUserEmail, RawlsUserSubjectId, UserInfo}
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
 
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.Await
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
@@ -17,7 +18,7 @@ class MarthaResolverSpec extends FlatSpecLike with Matchers with BeforeAndAfterA
   implicit val mockMaterializer: ActorMaterializer = ActorMaterializer()
   implicit val mockExecutionContext: TestExecutionContext = TestExecutionContext.testExecutionContext
 
-  val mockMarthaResolver = new MockMarthaResolver(marthaUrl = "https://martha_v2_url", excludeJDRDomain = true)
+  val mockMarthaResolver = new MockMarthaResolver(marthaUrl = "https://martha_v3_url")
   val mockUserInfo: UserInfo = UserInfo(RawlsUserEmail("mr_bean@gmail.com"), OAuth2BearerToken("foo"), 0, RawlsUserSubjectId("abc123"))
 
   behavior of "MarthaResolver"
@@ -41,29 +42,6 @@ class MarthaResolverSpec extends FlatSpecLike with Matchers with BeforeAndAfterA
 
     assertResult(Option("mr_bean@gmail.com")) {
       Await.result(actualResultFuture, 1 minute)
-    }
-  }
-}
-
-object MockMarthaResolver {
-  val jdrDevUrl = "drs://jade.datarepo-dev.broadinstitute.org/v1_0c86170e-312d-4b39-a0a4"
-  val dgUrl = "drs://dg.712C/fa640b0e-9779-452f-99a6-16d833d15bd0"
-
-  val mockEmail: ServiceAccountEmail = ServiceAccountEmail("mr_bean@gmail.com")
-  val mockSAPayload: ServiceAccountPayload = ServiceAccountPayload(Option(mockEmail))
-  val exampleGoogleSA: MarthaMinimalResponse = MarthaMinimalResponse(Option(mockSAPayload))
-}
-
-class MockMarthaResolver(marthaUrl: String, excludeJDRDomain: Boolean)
-                           (implicit system: ActorSystem, materializer: Materializer, executionContext: ExecutionContext)
-  extends MarthaResolver(marthaUrl) {
-
-  override def resolveDrsThroughMartha(drsUrl: String, userInfo: UserInfo): Future[MarthaMinimalResponse] = {
-    Future.successful{
-      drsUrl match {
-        case MockMarthaResolver.jdrDevUrl => MarthaMinimalResponse(None)
-        case _ => MockMarthaResolver.exampleGoogleSA
-      }
     }
   }
 }
